backend/requirements.txt
fastapi
uvicorn[standard]
pandas
pydantic
google-generativeai
python-dotenv